requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.5
passlib==1.7.4
psycopg2-binary==2.9.9
requests==2.31.0
PyJWT==2.8.0
python-dotenv
